docs: bruk Jøkul-komponenter for å bygge typografiskalaen
