--- portal/src/components/Documentation/Typography/TypograhyTable/TypographyTable.scss
@@ -1,63 +1,0 @@
-@use "~@fremtind/jkl-core/jkl";
-
-.jkl-typography-table {
-    border: 0;
-    width: 100%;
-    max-width: jkl.rem(768px);
-
-    &__data,
-    &__header {
-        padding: jkl.$spacing-xs;
-        @include jkl.text-style("body");
-        text-align: left;
-        vertical-align: top;
-        border-bottom: jkl.rem(1px) solid;
-        box-sizing: border-box;
-    }
-
-    &__header {
-        font-weight: jkl.$typography-weight-bold;
-        border-bottom-color: jkl.$color-granitt;
-        min-height: jkl.rem(48px);
-    }
-
-    &__data {
-        border-bottom-color: jkl.$color-stein;
-        padding-bottom: jkl.$spacing-m;
-        min-height: jkl.rem(64px);
-    }
-
-    @include jkl.small-device {
-        width: initial;
-
-        &__header {
-            display: none;
-        }
-
-        &__data {
-            display: flex;
-            border: none;
-            position: relative;
-            margin-left: jkl.rem(25px);
-            padding-left: jkl.rem(125px);
-
-            &:first-child {
-                padding-left: jkl.rem(20px);
-                margin-left: 0;
-            }
-
-            &::before {
-                @include jkl.text-style("small");
-                color: jkl.$color-stein;
-                content: attr(data-rowtitle);
-                position: absolute;
-                top: 13px;
-                left: 0;
-            }
-        }
-
-        &__row {
-            border-bottom: 1px solid jkl.$color-stein;
-        }
-    }
-}
