+++ portal/src/components/TypograhyTable/index.ts
@@ -1,0 +1,1 @@
+export { TypographyTable } from "./TypographyTable";
