+++ portal/src/texts/profile/typography.mdx
@@ -6,7 +6,7 @@ scss: webfonts
 order: 7
 ---
 
-import { TypographyExample } from "../../components/Documentation/Typography/TypograhyTable";
+import { TypographyTable } from "../../components/TypograhyTable";
 
 <Ingress>
     Typografi er et profilbærende element i løsningene vi lager. Vi bruker stor og tydelig typografi som grafisk
@@ -38,7 +38,7 @@ en skala.
     </InfoMessageBox>
 </aside>
 
-<TypographyExample />
+<TypographyTable />
 
 ## Overskrifter
